In BP_MotionController, fixed a bug setting the laser pointer scale - use the distance between the hit object and the origin of the laser pointer, instead of the absolute value of the X component.
In BP_MotionController, added a function "SetLaserPointerEnabled" to enable or disable the LaserPointer.

In BP_MotionController, changed the "Lighting -> Cast Shadow" option on the LaserPointer to false.

In MotionControllerPawn, added a section to the event graph responding to the input event "AnyKey". This section attempts to detect if any button is being pressed on either the left or right hand controller, and activate the appropriate laser pointer.

In BP_MotionController, added a widget interaction component, which is a child of the LaserPointer and has its interaction mode set to custom. Added functions that press and release the pointer key for the interaction component.

In BP_MotionController, added code to set the length of the laser pointer to the trace length (equal to the WidgetInteractionComponent's interaction distance) if the laser pointer is not hitting anything.

Added a trivial Widget to the level, for testing purposes.

In BP_MotionController, changed the "Line Trace By Profile" for the Laser Pointer to a "Line Trace By Channel", because I do not understand the former.